Move lib/fluent/plugin/file_wrapper.rb to lib/fluent
It doesn't use the name space `Fluent::Plugin`, and seems useful for
global usage rather than only for plugin.
