+++ lib/fluent/file_wrapper.rb
